Added an option to install dependencies through cmake without getlibs
