+++ src/hash/CMakeLists.txt
@@ -1,6 +1,6 @@
 file(GLOB hash_source "*.c" "*.cpp" "*.h")
 add_library(hash ${hash_source})
-target_link_libraries(hash PUBLIC xxHash::xxhash)
+target_link_libraries(hash PRIVATE xxHash::xxhash)
 
 
 target_include_directories(hash
