+++ test/base/test_random.cpp
@@ -1,7 +1,6 @@
 #include "base/random.h"
-
-#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include <doctest/doctest.h>
+#include <limits>
 
 auto rng = RandomGenerator{};
