+++ test/base/test_sequencefilter.cpp
@@ -7,12 +7,9 @@
 //
 ///////////////////////////////////////////////////////////////////////////////
 #include "base/sequence_filter.hpp"
-
-#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
+#include <doctest/doctest.h>
 #include <sstream>
 #include <vector>
-
-#include <doctest/doctest.h>
 
 using namespace std;
