--- test/io/CMakeLists.txt
@@ -1,1 +1,0 @@
-# Nothing
